+++ tools/add-lambda/data/handler.js
