+++ tools/add-lambda/data/package.json
