+++ tools/add-lambda/helpers/copy-files.js
@@ -1,0 +1,5 @@
+module.exports = (files, fs) => {
+  for (const file of files) {
+    fs.copyFileSync(file.src, file.dest)
+  }
+}
