+++ tools/add-lambda/helpers/create-folders.js
@@ -1,0 +1,5 @@
+module.exports = (folders, fs) => {
+  for (const folder of folders) {
+    fs.mkdirSync(folder)
+  }
+}
