+++ tools/add-lambda/helpers/get-args.js
@@ -1,0 +1,3 @@
+module.exports = (argsNames, args) => {
+  return Object.fromEntries(argsNames.map((argsName, index) => [argsName, args[index]]))
+}
