--- tools/add/helpers/copy-files.js
@@ -1,7 +1,0 @@
-const fs = require('fs')
-
-module.exports = (files) => {
-  files.forEach(file => {
-    fs.copyFileSync(file.src, file.dest)
-  })
-}
