--- tools/add/helpers/create-folders.js
@@ -1,7 +1,0 @@
-const fs = require('fs')
-
-module.exports = (folders) => {
-  folders.forEach(folder => {
-    fs.mkdirSync(folder)
-  })
-}
