--- tools/add/helpers/get-args.js
@@ -1,6 +1,0 @@
-module.exports = (argsNames, args) => {
-  const entries = argsNames.map((argsName, index) => {
-    return [argsName, args[index]]
-  })
-  return Object.fromEntries(entries)
-}
