--- tools/dev-server/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "devserver",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "koa": "^2.11.0",
-    "koa-better-router": "^2.1.1"
-  }
-}
