+++ tools/install.sh
@@ -1,14 +1,3 @@
-# Dev Server
-echo "installing dev-server"
-echo "changing directory"
-cd tools/dev-server
-echo $(pwd)
-echo "installing dependencies"
-npm i
-echo "done"
-echo "cleaning up"
-cd ../..
-echo $(pwd)
 # Layer
 echo "installing layer"
 echo "changing directory"
