+++ tools/update-version/update-version.js
@@ -1,12 +1,10 @@
 const fs = require('fs')
-const childProc = require('child_process')
-const updateApiVersion = require('./helpers/update-api-version.js')
-const updateLambdasVersion = require('./helpers/update-lambdas-version.js')
+const { spawnSync } = require('child_process')
 
 const args = process.argv.slice(2)
 const op = args[0] || 'patch'
-childProc.execSync(`npm version -no-git-tag-version ${op}`)
+spawnSync('npm', ['--no-git-tag-version', 'version', op], { stdio: 'inherit' })
 const version = JSON.parse(fs.readFileSync('package.json', 'utf8')).version
-updateApiVersion(version)
-updateLambdasVersion(version)
+require('./helpers/update-api-version.js')(version, fs)
+require('./helpers/update-lambdas-version.js')(version, fs)
 console.log(`Upgraded API & Lambdas version to ${version}`)
